refactor(libs): rename get-locale => locale-utils
