+++ deployer/aws-lambda/content-origin-request/package.json
@@ -17,7 +17,7 @@
   "dependencies": {
     "@yari-internal/constants": "file:../../../libs/constants",
     "@yari-internal/fundamental-redirects": "file:../../../libs/fundamental-redirects",
-    "@yari-internal/get-locale": "file:../../../libs/get-locale",
+    "@yari-internal/locale-utils": "file:../../../libs/locale-utils",
     "@yari-internal/slug-utils": "file:../../../libs/slug-utils",
     "accept-language-parser": "^1.5.0",
     "cookie": "0.5.0",
