+++ libs/locale-utils/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@yari-internal/get-locale",
+  "name": "@yari-internal/locale-utils",
   "version": "0.0.1",
   "private": true,
   "license": "MPL-2.0",
