+++ server/middlewares.ts
@@ -3,7 +3,7 @@ import express from "express";
 import { CSP_VALUE } from "../libs/constants";
 import { STATIC_ROOT } from "../libs/env";
 import { resolveFundamental } from "../libs/fundamental-redirects";
-import { getLocale } from "../libs/get-locale";
+import { getLocale } from "../libs/locale-utils";
 import { devMiddlewares } from "./dev";
 
 // Lowercase every request because every possible file we might have
